requirements.txt
pandas
selenium
webdriver-manager
openpyxl
Pillow